Added support for multiple point source lights.
Rewrote the LightFS fragment shader to hold an array of 4 lights. Changes propogated up through the relevant Javascript files.
Also modifed the source lights to attenuate over a distance instead of being a solid circle.

TODO: I'm uncomfortable with the size of this diff, need to look into doing some decoupling in the lighting engine.