Added spotlight and direction lighting logic to shaders.
Light and Illum shaders now respect spotlight and directional lighting. Illum shader appropriately applies specular and diffuse reflection to these new lighting sources.

Debugging this shit was impossible so I've also added some debug functionality to the shaders, highlighting the various transitional thresholds. For bugs in the future check: a) that everything is actually a unit vector (not a pixel vector that used to be a unit vector in WC space), b) check that all your vectors are pointing in the right direction, and c) that your shader logic is consistent.