GLSL shaders for shadow casting
Introducing two new shaders, ShadowCasterFS and ShadowReceiverFS.

Receiver is a simple one to one write of pixels. If the input texture is opaque enough, write a white pixel out. The output of the receiver shader will be piped through a stencil system in webGL instead of down the rest of the rendering pipeline. This only needs to be done once, after which all shadow casters' shadows will be drawn, appropriately sliced by the outlined stencil.
Caster is relatively simple, it draws a transluscent, tinted sillouette of the texture, which is then offset and scaled by logic in the engine.