Javascript Shader for new shadow casting system.
New caster that uses a single light and the object's texture to draw a transluscent tinted shadow. The shader will be swapped in for a Renderable's default shader, drawn at an appropriate offset and scale, then swapped back out. The shader will have to be run multiple times, once for each light the renderable is affected by, since the shadwo transformation isn't happening at the shader level.