Added LineRenderable, it draws lines
LineRenderable takes a start point, end point, and line width. Using these values and the color it inherits from Renderable, it creates a transform that draws our standard buffered square as a colored line with the appropriate width. Currently width is only in world units.

TODO: Add logic to set line with in pixels.