Fix LineRenderable rotation for certain end points
LineRenderable would render incorrectly when the line's end point's y value is less than that of the start value. The line appears as if it was flipped along the y axis. Fixed by appropriately negating the rotation value when the cross product's Z value is negative. Solution shamelessly ripped from GameObject's rotation logic.