Replace copied gl-matrix library with NPM module.
Added NPM build scripts to clean the build directory and compile the source code. Also added a postbuild hook that copies the minified gl-matrix source from the npm module to the build directory. Since the new source code exports in CommonJS (I think?), all references to gl-matrix provided definitions have to be preceded by "glMatrix.". This is ugly, but will be removed over the course of the Typescript migration so it's not BD Wong.