Convert RigidShapeTest to typescript.
I attempted to start the Typescript migration with this file three days ago but the issues I ran into during that work led me to refactor the entire project into Ecmascript modules. Now that everything is modularized, this conversion can finally  be completed. The missing camera methods are the only remaining issue, but will be resolved when Camera.js is converted.