Configure firefox debugger
Up until now I haven't been using VSCode's ability to attach to a browser debugger. This config file tells vscode how to attach to the firefox debugger when this project is running. Additional configuration is needed to streamline the process, currently the full procedure to attach to the debugger is in three steps:
- Ensure the live server is running, start it if not
- Build the code
- Attach to the firefox debugger using this new config

I'd like to see the server running be automatic eventually, but this is fine for now.