Remove two circular dependencies in engine core
The engine core module was responsible for two things: initializing the GL context, and initializing the rest of the engine modules.  Several of those modules depend on the GL context that core provided, causing a dependency loop.

The task of initializing and running the engine has been refactored into two function provided by Engine_Init (better name later). What was once Engine_Core is now Engine_GL and is solely responsible for providing the GL context.