Add parallax scrolling as a new GameObject
ParallaxGameObject extends TileGameObject to further modify that renderable's transform each step. The difference between the current position and last camera position is calculated. Then the length of the resulting vector is scaled up or down depending on how near or far, respectively, the object is to the camera, as defined by parallaxScale.