Move source code out of public_html
This diff aims to separate some top level folders to clarify their roles in the production process. Previously public_html was the folder intended to be served to clients for the webpage to run, but it contained the typescript source code.

src is now a top level folder and contains all typescript and glsl (this might change later) required to initialize, start and operate the game. The build process proceeds in the following way:
1) remove previously built (using rollup) js and sourcemap, as well as shaders.
2) use rollup which will delegate all typescript files to tsc to compile into jsvascript, then roll all the compiled javascript into a single output file. Place this file into public_html at the top.
3) copy shaders from src to public_html.