Fix hardcoded uniform location bug in SimpleShader
The globalAmbientIntensity property on SimpleShader was being set to 1.0 on construction, in an attempt to set the intensity to 1.0. This property's name is deceiving however; it holds the location of the intensity uniform location in the shader, not the intensity value being passed to the shader. By coincidence, the intensity uniform location was actually 1, so this was working fine by accident. Removing the default value assignment.