Introduce shader geometry types
Introduce a GeometryType enum with associated type data (num vertices per type, etc). Shader will hold one, and the user can swap it out to render a different shape. Under the hood, the shader is rebinding the appropriate vertex buffer (square or triangle) to the aShapeVertexPosition attribute on the shader.