Add method for checking switch case exhaustion
exhaustiveCheck should be called in the default case with a single argument, the object the switch statement is using to pick a case. If Typescript decides the method can never be reached, then the switch statement has exhausted all possible cases.

TODO: Lint rule to include this in switch statements.