Pass shape to shader instead of storing it there
Shaders were holding on to a shape property that GameObjects would set through the containing Renderable. (GameObject -> Renderable -> Shader)

Because shaders are reused, this led to a bug where renderables that didn't explicitly set their correct shape on the shader, would be rendered using whatever shape the shader had in hand.

By making it an argument to draw, the GameObject designer is expected to decide on a shape.