Tweak RigidRect to automatically detect dimensions
As title, before new instances of RigidRect had their dimensions passed to them. But since the rigid rect is holding on to its respective object's transform, this is usually redundant. It is still possible assign these dimensions manualyl, but now they default to the height and width of the provided xform.