Refactor collisions using the Strategy pattern
Collisions was a monolith class that held collision resolution logic for all combinations of RigidShapes. Even with 2 shapes and a Particle class, this is a mess. This is my first pass at refactoring this whole mess. I attempted to break the Collisions class down into multiple collision resolvers, one for each shape type.

The issue I'm running into now is deciding which resolver to use and that GameObjects don't have RigidShapes, they have PhysicsComponents.